backend/requirements.txt
# FastAPI Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Document Processing
python-docx>=1.1.0
ebooklib>=0.18
beautifulsoup4>=4.12.2
lxml>=5.0.0

# Audio/Video Processing
ffmpeg-python>=0.2.0

# Aliyun Services
dashscope>=1.14.1

# MinIO Storage
minio>=7.2.3

# HTTP Requests
requests>=2.31.0
httpx>=0.26.0

# Configuration Management
pydantic>=2.7.0
pydantic-settings>=2.2.0
python-dotenv==1.0.0

# Logging
loguru>=0.7.2

# Utilities
aiofiles>=23.2.1
